requirements.txt
pandas
numpy
scikit-learn
requests
wordcloud
matplotlib
seaborn
plotly
streamlit
pydantic

langchain-google-genai

# FIX DEPENDENCY (penting)
google-generativeai==0.8.6
google-ai-generativelanguage==0.6.15

# GCP & Sheets Integration
google-cloud-aiplatform
google-auth
gspread
python-dotenv
pyyaml